src/mcp-server/requirements.txt
mcp>=1.9.0
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.28.0
azure-identity>=1.19.0
